login/requirements.txt
boto3==1.19.8
PyJWT==2.3.0
pymongo==3.12.1